requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
psycopg2-binary==2.9.10

redis==5.0.4
boto3==1.34.122
python-dotenv==1.0.1
python-multipart==0.0.9

requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.2
pdfminer.six==20231228
